requirements.txt
PyPDF2
python-dotenv
streamlit 
google.generativeai
